PubGrub: move a few types to separate files
`PubGrubDependencyResolver.swift` reached almost 1.5k lines, making it hard to navigate in related areas of the codebase. Moving a few types from that file to separate files makes it smaller and more manageable. Additionally, `PubGrub` capitalization was made consistent in the naming.
